Add huge pages of 1G
For performance, closer to the paper config

+++ bootstrap.sh
@@ -10,28 +10,34 @@ sudo apt-get -y install python3 python3-pip build-essential meson pkg-config lib
                 libnet1-dev libyaml-0-2 libyaml-dev pkg-config zlib1g zlib1g-dev \
                 libcap-ng-dev libcap-ng0 make libmagic-dev         \
                 libnss3-dev libgeoip-dev liblua5.1-dev libhiredis-dev libevent-dev libjansson-dev liblz4-dev libpcre2-dev
-OFED=MLNX_OFED_LINUX-5.4-1.0.3.0-ubuntu${DISTRIB_RELEASE}-x86_64
-wget http://www.mellanox.com/downloads/ofed/MLNX_OFED-5.4-1.0.3.0/${OFED}.tgz
-tar xvf ${OFED}.tgz
-pushd ${OFED}
-sudo ./mlnxofedinstall --dpdk --upstream-libs --with-mft --with-kernel-mft --without-fw-update -q
-rm -f ${OFED}.tgz
 
-popd
-wget http://fast.dpdk.org/rel/dpdk-21.08.tar.xz
-tar xJf dpdk-21.08.tar.xz
-rm -f dpdk-21.08.tar.xz
-export DPDK_PATH=$(pwd)/dpdk-21.08
-export LD_LIBRARY_PATH=$DPDK_PATH/lib/x86_64-linux-gnu
-export PKG_CONFIG_PATH=$LD_LIBRARY_PATH/pkgconfig
+if [ ! -f OFED_INSTALLED ] ; then
+    OFED=MLNX_OFED_LINUX-5.4-1.0.3.0-ubuntu${DISTRIB_RELEASE}-x86_64
+    wget http://www.mellanox.com/downloads/ofed/MLNX_OFED-5.4-1.0.3.0/${OFED}.tgz
+    tar xvf ${OFED}.tgz
+    pushd ${OFED}
+    sudo ./mlnxofedinstall --dpdk --upstream-libs --with-mft --with-kernel-mft --without-fw-update -q
+    rm -f ${OFED}.tgz
+    popd
+    touch OFED_INSTALLED
+fi
 
-pushd $DPDK_PATH
-meson --prefix=$DPDK_PATH build
-cd build
-sudo ninja install
-sudo ldconfig
-ninja clean
-popd 
+if [ ! -f OFED_INSTALLED ] ; then
+    wget http://fast.dpdk.org/rel/dpdk-21.08.tar.xz
+    tar xJf dpdk-21.08.tar.xz
+    rm -f dpdk-21.08.tar.xz
+    export DPDK_PATH=$(pwd)/dpdk-21.08
+    export LD_LIBRARY_PATH=$DPDK_PATH/lib/x86_64-linux-gnu
+    export PKG_CONFIG_PATH=$LD_LIBRARY_PATH/pkgconfig
+
+    pushd $DPDK_PATH
+    meson --prefix=$DPDK_PATH build
+    cd build
+    sudo ninja install && touch ../DPDK_INSTALLED
+    sudo ldconfig
+    ninja clean
+    popd
+fi
 
 curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y 
 
@@ -49,18 +55,20 @@ pushd retina
 cargo build --release
 popd
 
-cargo install --force cbindgen
-git clone https://github.com/OISF/suricata.git
-pushd suricata
-#Build libhtp
-git clone https://github.com/OISF/libhtp
-pushd libhtp
-./autogen.sh && ./configure && make && sudo make install
-popd
-./autogen.sh
-./configure --enable-dpdk
-make
-sudo make install
-popd
+if [ ! -f suricata ] ; then
+    cargo install --force cbindgen
+    git clone https://github.com/OISF/suricata.git
+    pushd suricata
+    #Build libhtp
+    git clone https://github.com/OISF/libhtp
+    pushd libhtp
+    ./autogen.sh && ./configure && make && sudo make install
+    popd
+    ./autogen.sh
+    ./configure --enable-dpdk
+    make -j 16
+    sudo make install
+    popd
+fi
 
 echo "Bootstrap finished!"
